@@ -1,61 +1,46 @@
 class Api::ProductsController < ApplicationController
-    def index
-        # @products = Product.where(deleted_at: nil)
-        # @products = Product.all
-        @products = Product.all
-        render json: @products, include: [:category]
-    end
+  def index
+    @products = Product.all
+    render json: @products
+  end
 
-    def create   
-        # @products = Product.new(products_params)
-        # if @products.save
-        #     render json: @products 
-        # else
-        #     render json: { json: @products.errors, status: :unprocessable_entity }
-        # end
- 
-        @products = Product.new(products_params)
-        respond_to do |format|
-          if @products.save
-            params[:pictures]['title'].each do |a|
-               @picture = @products.pictures.create!(:title => a, :product_id => @products.id)
-            end
-            format.json { render json: @products.to_json }
-            format.html
-          else
-            format.html { render 'new'}
-            format.json { render json: @products.errors } 
-          end
-        end
+  def create
+    @products = Product.new(products_params)
+    if @products.save
+      InitServices::InitProduct.new(@products, params[:product][:pictures_ids]).perform
+      render json: @products
+    else
+      render status: 403
     end
+  end
 
-    def show
-        @products = Product.find(params[:id])
-        render json: { data: @products, status: :ok, message: 'Success' }
+  def show
+    @products = Product.find(params[:id])
+    render json: { data: @products, status: :ok, message: 'Success' }
+  end
+
+  def update
+    @products = Product.find(params[:id])
+    if @products&.update(products_params)
+      InitServices::InitProduct.new(@products, params[:product][:pictures_ids]).perform
+      render json: @products, status: 200
+    else
+      render json: { json: @products.error, status: :unprocessable_entity }
     end
+  end
 
-    def update
-        @products = Product.find(params[:id])
-        if @products.update(products_params)
-            render json: { status: :ok, message: 'Success' }
-        else
-            render json: { json: @products.error, status: :unprocessable_entity }
-        end
+  def destroy
+    @products = Product.find(params[:id])
+    if @products.destroy
+      render json: { json: 'Product was successfully deleted.' }
+    else
+      render json
     end
+  end
 
-    def destroy
-        @products = Product.find(params[:id])
-        if @products.destroy
-            render json: { json: 'Product was successfully deleted.'}
-        else
-            render json 
-        end
-    end
+  private
 
-    private
-    
-    def products_params
-        params.require(:product).permit(:name, :price, :description, :discount, :price_sale, :category_id, pictures_attributes: 
-            [:id, :product_id, :title])
-    end  
+  def products_params
+    params.require(:product).permit(:name, :price, :description, :discount, :price_sale, :category_id)
+  end
 end
