@@ -1,3 +1,3 @@
 class ApplicationController < ActionController::Base
-    protect_from_forgery unless: -> { request.format.json? }
+  protect_from_forgery unless: -> { request.format.json? }
 end
