@@ -1,4 +1,3 @@
 class WelcomeController < ApplicationController
-  def index
-  end
+  def index; end
 end
