@@ -46,6 +46,17 @@
 import newsModals from "../components/modals/news_modals";
 import axios from "axios";
 
+function dataURLtoFile(dataurl, filename) {
+  var arr = dataurl.split(","),
+    mime = arr[0].match(/:(.*?);/)[1],
+    bstr = atob(arr[1]),
+    n = bstr.length,
+    u8arr = new Uint8Array(n);
+  while (n--) {
+    u8arr[n] = bstr.charCodeAt(n);
+  }
+  return new File([u8arr], filename, { type: mime });
+}
 export default {
   data() {
     return {
@@ -126,25 +137,65 @@ export default {
             console.log(error);
           });
       } else {
-        let formData = new FormData();
-        formData.append("news[title]", item.title);
-        formData.append("news[description]", item.description);
-        formData.append("news[content]", item.content);
-        formData.append("news[image]", item.image.originFileObj);
-        axios
-          .put(`http://localhost:3000/api/news/${item.id}`, formData, {
-            headers: {
-              "Content-Type": "application/json"
-            }
-          })
-          .then(response => {
-            this.initialize();
-            this.$refs.child.close();
-            this.$message.success("Cập nhật bài viết thành công");
-          })
-          .catch(error => {
-            console.log(error);
+        if (item.image.url) {
+          let url = item.image.url;
+          const toDataURL = url =>
+            fetch(url)
+              .then(response => response.blob())
+              .then(
+                blob =>
+                  new Promise((resolve, reject) => {
+                    const reader = new FileReader();
+                    reader.onloadend = () => resolve(reader.result);
+                    reader.onerror = reject;
+                    reader.readAsDataURL(blob);
+                  })
+              );
+
+          toDataURL(url).then(dataUrl => {
+            let fileData = dataURLtoFile(dataUrl, item.title);
+
+            let formData = new FormData();
+            formData.append("news[title]", item.title);
+            formData.append("news[description]", item.description);
+            formData.append("news[content]", item.content);
+            formData.append("news[image]", fileData);
+            axios
+              .put(`http://localhost:3000/api/news/${item.id}`, formData, {
+                headers: {
+                  "Content-Type": "application/json"
+                }
+              })
+              .then(response => {
+                this.initialize();
+                this.$refs.child.close();
+                this.$message.success("Cập nhật bài viết thành công");
+              })
+              .catch(error => {
+                console.log(error);
+              });
           });
+        } else {
+          let formData = new FormData();
+          formData.append("news[title]", item.title);
+          formData.append("news[description]", item.description);
+          formData.append("news[content]", item.content);
+          formData.append("news[image]", item.image.originFileObj);
+          axios
+            .put(`http://localhost:3000/api/news/${item.id}`, formData, {
+              headers: {
+                "Content-Type": "application/json"
+              }
+            })
+            .then(response => {
+              this.initialize();
+              this.$refs.child.close();
+              this.$message.success("Cập nhật bài viết thành công");
+            })
+            .catch(error => {
+              console.log(error);
+            });
+        }
       }
     },
     editNews(item) {
