@@ -1,4 +1,4 @@
 class News < ApplicationRecord
-    # act_as_paranoid
-    mount_uploader :image, ImageUploader
+  # act_as_paranoid
+  mount_uploader :image, ImageUploader
 end
