@@ -1,4 +1,10 @@
 class Picture < ApplicationRecord
-  mount_uploader :title, ImageUploader
-  belongs_to :product
+  mount_uploader :image, ImageUploader
+  belongs_to :product, optional: true
+
+  def url
+    image.url
+  end
+
+  
 end
