@@ -11,15 +11,14 @@ class ImageUploader < CarrierWave::Uploader::Base
   # This is a sensible default for uploaders that are meant to be mounted:
   def store_dir
     "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
-      # "#{Rails.root}/app/javascript/assets/picture/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
-      
+    # "#{Rails.root}/app/javascript/assets/picture/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
   end
   # "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
 
   # def extension_white_list
   #   %w(jpg jpeg gif png)
   # end
- 
+
   # Provide a default URL as a default if there hasn't been a file uploaded:
   # def default_url(*args)
   #   # For Rails 3.1+ asset pipeline compatibility:
@@ -54,5 +53,4 @@ class ImageUploader < CarrierWave::Uploader::Base
   # def default_url
   #   "#{asset_host}/images/fallback/" + [version_name, "default.png"].compact.join('_')
   # end
-  
 end
