@@ -1,5 +1,5 @@
 # config/initializers/carrierwave.rb
 CarrierWave.configure do |config|
-    config.storage = :file
-    config.asset_host = ActionController::Base.asset_host
+  config.storage = :file
+  config.asset_host = ActionController::Base.asset_host
 end
