@@ -1,9 +1,8 @@
 require 'test_helper'
 
 class CategoryControllerTest < ActionDispatch::IntegrationTest
-  test "should get index" do
+  test 'should get index' do
     get category_index_url
     assert_response :success
   end
-
 end
