@@ -1,9 +1,8 @@
 require 'test_helper'
 
 class CustomersControllerTest < ActionDispatch::IntegrationTest
-  test "should get index" do
+  test 'should get index' do
     get customers_index_url
     assert_response :success
   end
-
 end
